#18576 merging indexresource and esindexresource
