+++ dotCMS/src/main/java/com/dotcms/rest/config/DotRestApplication.java
@@ -1,6 +1,5 @@
 package com.dotcms.rest.config;
 
-import com.dotcms.rest.api.v1.pushpublish.PushPublishFilterResource;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
@@ -36,6 +35,7 @@ import com.dotcms.rest.api.v1.page.PageResource;
 import com.dotcms.rest.api.v1.personalization.PersonalizationResource;
 import com.dotcms.rest.api.v1.personas.PersonaResource;
 import com.dotcms.rest.api.v1.portlet.PortletResource;
+import com.dotcms.rest.api.v1.pushpublish.PushPublishFilterResource;
 import com.dotcms.rest.api.v1.relationships.RelationshipsResource;
 import com.dotcms.rest.api.v1.site.SiteResource;
 import com.dotcms.rest.api.v1.sites.ruleengine.rules.RuleResource;
@@ -57,7 +57,6 @@ import com.dotcms.rest.api.v1.user.UserResource;
 import com.dotcms.rest.api.v1.vtl.VTLResource;
 import com.dotcms.rest.personas.PersonasResourcePortlet;
 import com.dotcms.rest.servlet.ReloadableServletContainer;
-import com.dotmarketing.common.reindex.IndexResource;
 import com.google.common.collect.ImmutableSet;
 
 /**
@@ -154,7 +153,6 @@ public class DotRestApplication extends javax.ws.rs.core.Application {
 			.add(BrowserResource.class)
 			.add(ResourceLinkResource.class)
 			.add(PushPublishFilterResource.class)
-	        .add(IndexResource.class)
 			.build();
