+++ dotCMS/src/main/java/com/dotmarketing/portlets/cmsmaintenance/ajax/IndexAjaxAction.java
@@ -20,6 +20,7 @@ import com.dotmarketing.util.UtilMethods;
 import com.dotmarketing.util.WebKeys;
 import com.liferay.portal.language.LanguageUtil;
 import com.liferay.portal.model.User;
+import io.vavr.control.Try;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
@@ -27,6 +28,7 @@ import java.io.OutputStream;
 import java.io.PrintWriter;
 import java.nio.file.Files;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 import java.util.Map;
 import javax.servlet.ServletException;
@@ -121,98 +123,21 @@ public class IndexAjaxAction extends AjaxAction {
 
 	}
 
-	public void restoreIndex(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException, DotDataException {
-	    try {
-            FileItemFactory factory = new DiskFileItemFactory();
-            ServletFileUpload upload = new ServletFileUpload(factory);
-            List<FileItem> items = (List<FileItem>) upload.parseRequest(request);
-
-            String indexToRestore=null;
-            boolean clearBeforeRestore=false;
-            String aliasToRestore=null;
-            File ufile=null;
-            boolean isFlash=false;
-            for(FileItem it : items) {
-               if(it.getFieldName().equalsIgnoreCase("indexToRestore")) {
-                   indexToRestore=it.getString().trim();
-               }
-               else if(it.getFieldName().equalsIgnoreCase("aliasToRestore")) {
-                   aliasToRestore=it.getString().trim();
-               }
-               else if(it.getFieldName().equalsIgnoreCase("uploadedfiles[]") || it.getFieldName().equals("uploadedfile")
-                       || it.getFieldName().equalsIgnoreCase("uploadedfileFlash")) {
-                   isFlash=it.getFieldName().equalsIgnoreCase("uploadedfileFlash");
-                   ufile=File.createTempFile("indexToRestore", "idx");
-                   InputStream in=it.getInputStream();
-                   final OutputStream out = Files.newOutputStream(ufile.toPath());
-                   IOUtils.copyLarge(in, out);
-                   IOUtils.closeQuietly(out);
-                   IOUtils.closeQuietly(in);
-               }
-               else if(it.getFieldName().equalsIgnoreCase("clearBeforeRestore")) {
-                   clearBeforeRestore=true;
-               }
-            }
-
-            if(ufile!=null) {
-                ESIndexResource.restoreIndex(ufile, aliasToRestore, indexToRestore, clearBeforeRestore);
-            }
-            
-            PrintWriter out=response.getWriter();
-            if(isFlash) {
-                out.println("response=ok");
-            }
-            else {
-                response.setContentType("application/json");
-                out.println("{\"response\":1}");
-            }
-            
-	    }
-	    catch(FileUploadException fue) {
-	        Logger.error(this, "Error uploading file", fue);
-	        throw new IOException(fue);
-	    }
-	}
-
 	
-	public void downloadIndex(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException, DotDataException {
-		Map<String, String> map = getURIParams();
-		response.setContentType("application/zip");
 
-		String indexName = indexHelper.getIndexNameOrAlias(map,"indexName","indexAlias",this.indexAPI);
-		if(!UtilMethods.isSet(indexName)) return;
-		
-		File f=ESIndexResource.downloadIndex(indexName);
-		response.setContentLength((int) f.length());
-		OutputStream out = response.getOutputStream();
-		final InputStream in = Files.newInputStream(f.toPath());
-
-		response.setHeader("Content-Type", "application/zip");
-		response.setHeader("Content-Disposition", "attachment; filename=" + indexName + ".zip");
-
-		IOUtils.copyLarge(in, out);
-
-		f.delete();
-	}
 
 	public void createIndex(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException, DotIndexException {
 
 		Map<String, String> map = getURIParams();
-		int shards = 0;
-
-		try{
-			shards = Integer.parseInt(map.get("shards"));
-
-		}
-		catch(Exception e){
-
-		}
+		final int shards = Try.of(()->Integer.parseInt(map.get("shards"))).getOrElse(1);
 
 
-		boolean live = map.get("live") != null;
-		String indexName = map.get("indexName");
-		
-		ESIndexResource.create(indexName, shards, live);
+		final boolean live = map.get("live") != null;
+		final String indexName=((live) ? "live_" : "working_" ) + APILocator.getContentletIndexAPI().timestampFormatter.format(new Date());
+
+
+
+		APILocator.getContentletIndexAPI().createContentIndex(indexName, shards);
 	}
 
 	public void clearIndex(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException, DotStateException, DotDataException {
@@ -236,13 +161,13 @@ public class IndexAjaxAction extends AjaxAction {
 		Map<String, String> map = getURIParams();
 		String indexName = indexHelper.getIndexNameOrAlias(map,"indexName","indexAlias",this.indexAPI);
 
-		ESIndexResource.activateIndex(indexName);
+		APILocator.getContentletIndexAPI().activateIndex(indexName);
 
 	}
 	public void deactivateIndex(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException, DotDataException {
 		Map<String, String> map = getURIParams();
 		String indexName = indexHelper.getIndexNameOrAlias(map,"indexName","indexAlias",this.indexAPI);
-		ESIndexResource.deactivateIndex(indexName);
+		APILocator.getContentletIndexAPI().deactivateIndex(indexName);
 	}
 
 	@Override
@@ -330,12 +255,6 @@ public class IndexAjaxAction extends AjaxAction {
 		response.getWriter().println(resp);
     }
 
-	public void getIndexRecordCount(HttpServletRequest request, HttpServletResponse response) throws IOException {
-		Map<String, String> map = getURIParams();
-		String indexName = indexHelper.getIndexNameOrAlias(map,"indexName","indexAlias",this.indexAPI);
-
-		response.getWriter().println(ESIndexResource.indexDocumentCount(indexName));
-	}
 
 	public void getNotActiveIndexNames(HttpServletRequest request, HttpServletResponse response) throws IOException {
 		ContentletIndexAPI idxApi = APILocator.getContentletIndexAPI();
